reactivate tests
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/vplants/vplants/trunk/PlantGL@5968 ab253cce-29fb-0310-bb2f-979600cdbdeb

+++ test/test_viewer.py
@@ -2,16 +2,11 @@ from openalea.plantgl.all import *
 from math import *
 import os
 
-from PyQt4.QtCore import *
-from PyQt4.QtGui import *
-
-import time
-
-app = QApplication([])
-
-
-
 def test_start():
+    from PyQt4.QtCore import QCoreApplication
+    import warnings
+    if not QCoreApplication.instance() is None:
+        warnings.warn("A QApplication is already running")
     Viewer.start()
   
 def test_display():
@@ -109,32 +104,26 @@ def test_camera_light():
     assert Viewer.position == pos, "Viewer.light.position do not set the good value"
  
 
-# This test needs to be fixed. It is not robust.
-# For instance, w and h are not equql to rw qnd rh respectively
-# Moreover, it fails in buildbot
-# to put back this test change its name to test_image 
 def test_image():
     w = 400
     h = 400
-    Viewer.frameGL.maximize()
+    rw, rh = w,h
+    Viewer.widgetGeometry.setSize(w,h)
+    Viewer.frameGL.maximize(True)
     Viewer.frameGL.setSize(w, h)
+    Viewer.update()
     fname = 'test_framegl.png'
     Viewer.frameGL.saveImage(fname,'PNG')
     Viewer.frameGL.maximize(False)
     assert os.path.exists(fname), "Viewer.frameGL.saveImage failed"
-    #try:
     if True:
         from PIL import Image
         q = Image.open(fname)
         rw, rh = q.size[0], q.size[1]
-        # the final dimensions (rw,rh) are larger than requested dimensions (r,w)
-        # the excess is constant (172,108). Is it the expected behaviour ?
-        imgsizetest = (rw == (w+172) and rh == (h+108))
+        imgsizetest = (rw == w) and (rh == h)
         del q
-    #except:
-    #    imgsizetest = True
-
-    assert imgsizetest and "Viewer.frameGL.setSize failed %s" % str((rw,rh))
+    if not imgsizetest :
+        raise Exception( "Viewer.frameGL.setSize failed %s" % str((rw,rh)))
     os.remove(fname)
 
 def test_state():
@@ -159,22 +148,13 @@ if __name__=='__main__':
         print 'Avoid display.'
     except:
         test_start()
-        time.sleep(1)
         test_display()
-        time.sleep(1)
         test_add()
-        time.sleep(1)
         test_update()
-        time.sleep(1)
         test_selection()
-        time.sleep(1)
         test_scene_interaction()
-        time.sleep(1)
         test_camera_lookat()
-        time.sleep(1)
         test_camera_set()
-        time.sleep(1)
         test_camera_light()
         test_image()
-        time.sleep(1)
         test_state()
